+++ FlappyBirdAI/util/MathUtil.h
@@ -4,7 +4,7 @@
 // fit parabola by x[], y[]
 // y = ax^2 + bx + c
 void ParabolaFit(
-    const std::vector<float>& x,
+    const std::vector<double>& x,
     const std::vector<float>& y,
     /*OUT*/ double& a,
     /*OUT*/ double& b,
